# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Tamra application",
        "title": "Tamra API",
        "contact": {},
        "version": "1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/restaurants": {
            "post": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Create a new restaurant with the given request body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Create a new restaurant",
                "parameters": [
                    {
                        "description": "Create Restaurant Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateRestaurantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Restaurant",
                        "schema": {
                            "$ref": "#/definitions/models.Restaurant"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create restaurant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restaurants/logo/uploadurl": {
            "get": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Get a signed URL to upload a restaurant logo to the S3 bucket",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Get a signed URL to upload a restaurant logo",
                "responses": {
                    "200": {
                        "description": "Presigned URL",
                        "schema": {
                            "$ref": "#/definitions/models.RestaurantLogoUploadResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get upload URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/restaurants/me": {
            "get": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Get a restaurant by the user ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Get a restaurant",
                "responses": {
                    "200": {
                        "description": "Restaurant",
                        "schema": {
                            "$ref": "#/definitions/models.Restaurant"
                        }
                    },
                    "404": {
                        "description": "Restaurant not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to get restaurant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "description": "Update a restaurant with the given request body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Update a restaurant",
                "parameters": [
                    {
                        "description": "Update Restaurant Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateRestaurantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated Restaurant",
                        "schema": {
                            "$ref": "#/definitions/models.Restaurant"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update restaurant",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateRestaurantRequest": {
            "type": "object",
            "required": [
                "latitude",
                "logo_url",
                "longitude",
                "name"
            ],
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "logo_url": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Restaurant": {
            "type": "object",
            "required": [
                "latitude",
                "logo_url",
                "longitude",
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "logo_url": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.RestaurantLogoUploadResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "presigned_url": {
                    "type": "string"
                },
                "stored_file_url": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateRestaurantRequest": {
            "type": "object",
            "required": [
                "latitude",
                "logo_url",
                "longitude",
                "name"
            ],
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "logo_url": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "jwt": {
            "description": "Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}